Add a leaderboard to the in game screen
Currently only based on the data stored in current browser's localStorage

It is drawn once at the start of the game and then gets updated anytime the current game's score updates (end of level).

The scores are extracted from saved profile but in a roundabout way, we
can probably simplify this by modifying the Ruby code to read a base64
encoded profile directly instead of from a file.