Only allow a unique name per playthrough
If we don't guard against this, it's possible to resume a game. But for
ease of leaderboard management (both elapsed times and getting your profile
hijacked), we will only allow one name.

For ease of leaderboard tracking, we will normalize names used in
localstorage keys as uppercase. The warrior name passed to the ruby
profile will keep its casing intact (and thus be reflected in the
leaderboard)